services/ocr/requirements.txt
fastapi
uvicorn
python-multipart
pytesseract
Pillow
uuid
python-dotenv
